Event: split out model
